+++ container-engine-lib/docs/changelog.md
@@ -1,6 +1,7 @@
 # TBD
 ### Features
 * Added `RemoveVolume` and `RemoveContainer` functions
+* Added a `GetVolumesByName` function
 
 ### Changes
 * Clarified that the `all` argument to `GetContainersByLabels` is for whether stopped containers should be shown